pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rekordbox-edit"
version = "0.3.0"
description = "Tools for managing and modifying a RekordBox library en-masse"
authors = [{ name = "James Viall", email= "jamesviall@pm.me"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["rekordbox", "dj", "music", "database", "bulk", "edit", "convert"]
dependencies = [
    "pyrekordbox==0.4.4",
    "ffmpeg-python>=0.2.0",
    "click>=8.0.0,<=9.0.0",
    "platformdirs>=4.3.8,<5.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1,<9",
    "commitizen>=4.8.3,<5",
    "pre-commit>=4.2.0,<5",
    "ruff>=0.12.8,<1",
    "pytest-cov>=6.2.1,<7",
    "pytest-mock>=3.14.1,<4",
    "callee>=0.3.1,<1",
    "ty>=0.0.1,<1",
]

[project.urls]
Homepage = "https://github.com/jviall/rekordbox-edit"
Repository = "https://github.com/jviall/rekordbox-edit"
Issues = "https://github.com/jviall/rekordbox-edit/issues"

[project.scripts]
rekordbox-edit = "rekordbox_edit.cli:main"
rbe = "rekordbox_edit.cli:main"

[tool.coverage.run]
branch = true
data_file = ".coverage/coverage.db"
[tool.coverage.html]
directory = ".coverage/html/"
[tool.coverage.xml]
output = ".coverage/coverage.xml"
[tool.coverage.json]
output = ".coverage/coverage.json"
[tool.coverage.lcov]
output = ".coverage/coverage.lcov"

[tool.ty.environment]
python-version = "3.10"

[tool.ty.analysis]
allowed-unresolved-imports = ["pyrekordbox", "ffmpeg"]

[tool.commitizen]
name = "cz_customize"
tag_format = "v$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true

[tool.commitizen.customize]
message_template = "{{change_type}}{{scope}}: {{message}}"
example = "feat(convert): add new conversion feature"
schema = "<type>(<scope>): <body>"
schema_pattern = "^(BREAKING|feat|fix|refactor|ci|test|docs|chore)(\\([^)]*\\))?!?: .+"
bump_pattern = "^(BREAKING|feat|fix|perf|hotfix|refactor)"
bump_map = { "BREAKING" = "MAJOR", "feat" = "MINOR", "fix" = "PATCH", "refactor" = "PATCH" }

[[tool.commitizen.customize.questions]]
type = "list"
name = "change_type"
choices = [
    {value = "BREAKING", name = "BREAKING: introduces breaking changes"},
    {value = "feat", name = "feat: introduces new features"},
    {value = "fix", name = "fix: patches a bug"},
    {value = "refactor", name = "refactor: code refactor with no functionality changes"},
    {value = "ci", name = "ci: only CI/CD changes"},
    {value = "test", name = "test: only adds or modifies tests"},
    {value = "docs", name = "docs: only documentation changes"},
    {value = "chore", name = "chore: some other non-code changes such as configs or dev dependency updates"}
]
message = "Select the type of change you are committing"

[[tool.commitizen.customize.questions]]
type = "input"
name = "scope"
message = "What is the scope of this change? (convert/read) (press [enter] to skip)"

[[tool.commitizen.customize.questions]]
type = "input"
name = "message"
message = "Write a short and imperative summary of the code changes"
